Bugfix/weilei/cicd (#205)
* add log

* mv resource of cicd to workload env

* mv cicd resource to workload.env

* upgrade copyright

* format code

* change proxy image name

* use test-runner for primus-safe

* format code

* update metadata to k8s object

* rollback build.yml

* fix after review

+++ SaFE/apiserver/pkg/handlers/custom-handlers/workspace.go
@@ -540,7 +540,7 @@ func (h *Handler) cvtToGetWorkspaceResponse(ctx context.Context, workspace *v1.W
 	}
 	nfResource := nf.ToResourceList(commonconfig.GetRdmaName())
 
-	abnormalQuota := quantity.MultiResource(nfResource, int64(result.AbnormalNodeCount))
+	abnormalQuota := quantity.MultiResource(nfResource, float64(result.AbnormalNodeCount))
 	result.TotalQuota = cvtToResourceList(workspace.Status.TotalResources)
 	result.AbnormalQuota = cvtToResourceList(abnormalQuota)
